Changed formal context cell size to 52-bit. There is problem with JS bitwise operators, that can work with 32-bit integers only

+++ src/cpp/conceptsCover.cpp
@@ -13,11 +13,12 @@
 #include <memory>
 #include <unordered_set>
 #include <map>
+#include <cstdint>
 
 void conceptsCover(
     TimedResult<std::vector<std::vector<int>>>& result,
     std::vector<SimpleFormalConcept>& concepts,
-    std::vector<unsigned int>& contextMatrix,
+    std::vector<uint64_t>& contextMatrix,
     int cellSize,
     int cellsPerObject,
     int contextObjectsCount,
